Fixed printf format errors.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@37239 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/newfs/newfs.c
@@ -541,7 +541,7 @@ havelabel:
 	if (t_or_u_flag && secpercyl != lp->d_secpercyl)
 		fprintf(stderr, "%s (%d) %s (%lu)\n",
 			"Warning: calculated sectors per cylinder", secpercyl,
-			"disagrees with disk label", lp->d_secpercyl);
+			"disagrees with disk label", (u_long)lp->d_secpercyl);
 	if (maxbpg == 0)
 		maxbpg = MAXBLKPG(bsize);
 	headswitch = lp->d_headswitch;
